ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
sentence-transformers==3.3.1
scikit-learn==1.6.1
numpy==2.2.1
nltk==3.9.1
requests==2.32.3
aiofiles==24.1.0
